usb: dwc3: core: make USB3 PHY optional
If we want a port to work at any speed lower than Superspeed, it makes no
sense to even initialize/power up the USB3 transceiver, provided it won't
be used. We can use the oportunity to save some power and leave the
superspeed transceiver powered off. There is at least one such case which
is Texas Instruments' AM437x which has one of its USB3 ports without a matching
USB3 PHY (that port is hardwired to work on USB2 only).

http://git.kernel.org/cgit/linux/kernel/git/balbi/usb.git/commit/
?h=testing&id=d7e39d414310e098540605be2051d5187797be34

Change-Id: I61639c4d58aac671a5aade67814981dc6d20f2e6
Signed-off-by: Felipe Balbi <balbi@ti.com>
Signed-off-by: Ruchika Kharwar <ruchika@ti.com>