HACK: usb: dwc3: dwc3 wrapper update for multiple usb subsystems
Allow a separate dma mask per usb_otg_ss instance.
REVISIT: This is based on the name of the node in the device tree and ugly.
Temporary to get DRA7xx going and hardly scalable.

"_omap" global still exists and the only reason his works is because dra7xx
does not use the palmas driver.

Change-Id: Iff9d09e837c1556516c28c42f7a3f41a2315befc
Signed-off-by: Ruchika Kharwar <ruchika@ti.com>