clk: omap5: Add CLK_SET_RATE_PARENT flag to gpu clocks
clk_set_rate finds the topmost clock that needs to be changed
in order to set the desired clock to a particular frequency.

This is done only if CLK_SET_RATE_PARENT flag is set for a clock.
gpu_core_gclk_mux and gpu_hyd_gclk_mux need to have this flag
enabled in order for clk_set_rate to propagate to to their parent.

Signed-off-by: Hemant Hariyani <hemanthariyani@ti.com>