ARM: DRA7: hwmod: HDMI in SWSUP while active
DRA7XX TRM states that no-idle mode is recommended to be used in case
audio is transferred (to avoid interface clock to be shut down).
There are additional PM framework changes needed to implement this
constraint dynamically on audio activity basis, so using SWSUP only
when the HDMI module is active is better option at the moment.

Change-Id: Id6f39e25a39c544679146ca1f3118289b56ee968
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>