HACK: rpmsg: rpc: avoid parameter packet corruption
The rpmsg-rpc .write system call imports the necessary dma-buf
pointers at run-time (if they have not been registered before)
during a function invocation call.

For some unknown reason, the rpmsg-rpc write system call is
getting executed twice sometimes, with the buffer lookup for the
parameter structure executed twice. During the first iteration,
dmabuf has been successfully imported for the param buffer fd
and is added to the idr list. However during the second iteration,
the dma buffer is found in the idr list, and the dma buffer import
is skipped, sending the incomplete/corrupted dma buffer to the
remote core. This results in a crash on the remote core.

The current fix is a temporary HACK to handle or ignore the dma
buffers that were already imported for the same fd. This patch
will be reverted once the issue is rootcaused and fixed.

Change-Id: Ie2027546bd53717b4912f3645ece49f4e0bc593d
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>