OMAPDSS: Add methods for skipping display initialization
Adds methods to handle skipping display initialization during
kernel bootup for cases where the display is pre-initialized by
either the remote processor or u-boot.

The skip_init flag is set from the dts file. This allows the
init code to be skipped. If this is initialized to 0, the kernel
does all the usual display initialization procedures.

Change-Id: I91d3e10cdb9ae5a2b8a46ba6f4b1eb82acead702
Signed-off-by: Sundar Raman <sunds@ti.com>