OMAPDSS: Adapt DSS driver for early camera use-case
This patch adds support for early camera use-case in
dss. The remote processor uses VID3_ENDWINDOW interrupt
for frame completion whereas the kernel dss driver uses
VSYNC. Additional changes are done to the dss driver
to ensure that the remote processor's irq flags are not
reset inadvertently.

A new config CONFIG_EARLYCAMERA is introduced and all
changes are made under it.

Change-Id: Ie53b632ea695dde8bf1d4ed109b443bbee24c6d0
Signed-off-by: Sundar Raman <sunds@ti.com>