drivers: net: cpsw: fix cpsw_ale_destroy
When cpsw is used as a module, and the ethernet device is
not opened (for eg: dhcpcd is not running yet),
there is a deadlock if the module is unloaded (delete
the timer). This is because the destory function tries
to delete a timer thats not initialized. The timer is
stopped in cpsw_ndo_stop, and hence cpsw_ale_stop need
not be called in destroy.

Change-Id: Ib815f19a7f8720355b552319323026a3b1c762de
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>
Reviewed-by: Mugunthan V N <mugunthanvnm@ti.com>