C_CAN: D_CAN: Remove non-existent DCAN RAM de-init
For DCAN support, a method for CAN RAM init was added. A parameter for
this method was included to specify initialization or de-initialization
of the message RAM. However, there is no hardware support for RAM
de-initialization, nor any need to repeatedly call this invalid de-init
sequence. Thus, remove the parameter in all definitions of the method
and remove all de-init calls to the method.

Change-Id: I5e653e1c9030b1a566b5932ec208b4f853217c2a
Signed-off-by: Austin Beam <austinbeam@ti.com>