C_CAN: D_CAN: Add explicit RAMINIT bit containers
Allow disassociation of RAMINIT bits from instance ID. As-is, the
RAMINIT bits are assumed to map 1:1 with the instance ID of the DCAN
instance. With this addition, RAMINIT bits are defined by default using
this assumption, but can be modified if needed. The new raminit_bits
structure should be used in place of the instance ID for RAMINIT bit
positions going forward.

Change-Id: I3366c5db462f9a70feba6dcb30c2fbf011b25fae
Signed-off-by: Austin Beam <austinbeam@ti.com>