HACK: C_CAN: D_CAN: DRA7: Disable IRQs in RAMINIT
This hack disables IRQs while the RAMINIT sequence occurs to address an
errata on DRA7xx ES1.0 which requires that the RAMINIT_START bit be set
and cleared within 64 OCP clock cycles. By disabling IRQs, the potential
is reduced that this deadline is missed. This change is not needed for
DRA7 >ES1.0.

Change-Id: I78d48cf5276de72e695c7aa55024c92ff97dfe28
Signed-off-by: Austin Beam <austinbeam@ti.com>