ARM: dts: dra7-evm: Add flag to skip display initialization
Add a separate dts flag to handle skipping display initialization during
kernel bootup for cases where the display is pre-initialized by
either the remote processor or u-boot.

The skip_init flag is set from the dts file. This allows the
init code to be skipped. If this is initialized to 0, the kernel
does all the usual display initialization procedures.

Change-Id: I19f1191539d50dfedee80ca4e4536ba73e722502
Signed-off-by: Sundar Raman <sunds@ti.com>