OMAPDSS: LCD Panel: Retain GPMC-FPDLINK GPIO to high in panel probe
The gpio pin used in this panel via a pcf gpio expander
is used to select between GPMC and VOUT3.

GPIO PIN	SELECTION
-------------------------
HIGH		GPMC INTERFACE
LOW		VOUT3

Since the default setting should select GPMC, set the
GPIO to be in default value (HIGH).

It is required to have GPMC selected since ROM code hangs
after soft reset. Currently we have the GPIO pulled low causing
the softreset to fail due to hang.

To temporary overcome this issue when FPDLINK panel is not
connected we keep the GPIO as HIGH.

WHEN FPDLINK IS CONNECTED WE ARE EXPECTING THE SOFTRESET TO HANG

Change-Id: I9a121e328206e4d6b94f4302f3537696afe1a1eb
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>
Signed-off-by: Dandawate Saket <dsaket@ti.com>