PANEL: Fix to remove the conflict between HDMI/VIP
This is a valid fix to remove the conflict that
existed between HDMI and VIP. HDMI takes precedence
over VIP since they can not be functional at the
same time.

When we receive an interrupt signaling that HPD
is low (HPD 0), we then precede to entering the
hdmi_hotplug_detect_worker thread. Since the HPD
state is off the worker thread then checks if
there is an active display, but since there is not
the state machine must be restored.

Change-Id: I9477129199a02346d56b5431cf6b535f248b7d3b
Signed-off-by: Arthur Philpott <arthur.philpott@ti.com>