C_CAN: D_CAN: Restore DCAN RAM de-init
DCAN RAMINIT previously supported de-init, but this feature was removed
with change I5e653e1c9030b1a566b5932ec208b4f853217c2a. The feature was
removed because the device control module is supposed to automatically
handle restoration of the DCAN initialization start bit to 0. However,
it is better to explicitly return the start bit to 0 and this feature
should not have been removed.

Thus, the following is accomplished with this change:
- Revert "C_CAN: D_CAN: Remove non-existent DCAN RAM de-init"
- Add support for de-init to the dra7-specific DCAN RAM init method

Change-Id: I5ccf909048bce88982d835fcaeffe48ca3644b87
Signed-off-by: Austin Beam <austinbeam@ti.com>