HACK: DRA7: DT: Fix BT regression due to DCAN1 clk
Adding dcan1_sys_clk_mux caused a BT regression where BT failed to
initialize properly. After investigation, it became clear that the
dcan1_sys_clk_mux initialization is not required for DCAN1 to operate
properly. By providing sys_clkin1 as the DCAN1 clock, the DCAN1 module
is functional. Furthermore, by removing dcan1_sys_clk_mux from DT, the
BT regression is resolved.

This is considered a HACK because it is not fully understood why this is
impactful to BT. Furthermore, it is not fully understood why the
dcan1_sys_clk_mux is not required to be enabled for DCAN1 functionality.

Change-Id: I9aeb37a278c1d8f776cabf37ff39ad666f07c501
Signed-off-by: Austin Beam <austinbeam@ti.com>