v4l: ti-vpe: make sure VPDMA line stride constraints are met
When VPDMA fetches or writes to an image buffer, the line stride must be a
multiple of 16 bytes. If it isn't, VPDMA HW will write/fetch until the next
16 byte boundry.

In order to prevent this, we now make sure that the VPE source and destination
image line strides are 16 byte aligned. Also, the motion vector buffer must be
allocated in such a way that it ensures the same alignment.

Change-Id: Iff0c3a433ecfeb73a4cdda9a50229d5d3c6c4f4b
Signed-off-by: alaganraj <alaganraj.s@ti.com>