v4l: ti-vpe: enable YUV to RGB color conversion support
The CSC block can be used for color space conversion between YUV and RGB
formats.

It is configurable via a programmable set of coefficients. Add functionality to
choose the appropriate CSC coefficients and program them in the CSC registers.
We take the source and destination colorspace formats as the arguments, and
choose the coefficient table accordingly.

Only YUV to RGB coefficients are provided for standard and high definition
colorspaces. The coefficients can also be limited or full range, only full range
coefficients are chosen by default. We would need some sort of control ioctl for
the user to specify the range needed.

Change-Id: I1aa76dfcc0702ef022e8516f59a64f953f7331c5
Signed-off-by: alaganraj <alaganraj.s@ti.com>