v4l: ti-vpe: Create a scaler library
Create the scaler block related configurations and register definitions as a
library. This is done as the same scaler block is used in both VIP and VPE.

The vpe_dev device holds a sc_data handle, which it can reference to access
the registers, or call scaler related helper functions.

Change-Id: Ia785ba243938759555dd7bca9e6f547be00c07db
Signed-off-by: alaganraj <alaganraj.s@ti.com>

Conflicts:

	drivers/media/platform/Makefile

Signed-off-by: Praneeth Bajjuri <praneeth@ti.com>