v4l: ti-vpe: Add sync on channel descriptors for input VPE ports
We currently stall the VPDMA descriptor list till the time the output channels
complete their DMA. In certain cases, this may not be the correct completion
event. It's possible that there is some pending DMA operation still occuring
on one of the input channels.

Add sync on channel control descriptors for the input channels to make sure
a list complete interrupt occurs when all the channels are done with DMA.

Change-Id: I24d3127bf854fe8ac145e33803665a7a58a1d042
Signed-off-by: alaganraj <alaganraj.s@ti.com>