ASoC: tlv320aic3x: Fix incorrect locking order
The .put() function used for a subset of Left and Right PGA
Mixers was acquiring the ASoC locks in an incorrect order:
codec's mutex and then DAPM mutex.

The involved widgets/controls are mixers, so they require
DAPM functions to update routes that are internal to ASoC.
The available functions are not feasible as they are locked
versions of the ones needed to implement the .put() callback.
Unlocked versions of the power update for mixers and muxes are
exported now.

The implementation of the .put() callback has also been updated
to keep it more in sync with the default one for that type of
widgets.

Change-Id: Id031c3fa49aa0130f0dd105f4d5afcb16c68e3d7
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>