ARM: OMAP: remoteproc: adjust the CMA pool size for OMAP5 IPU
The CMA pool size for IPU on OMAP5 is reduced from the previous
168 MB to 56 MB. This reduction is done to account for the removal
of the OMAP camera stack, and to match to that of the size used
on DRA7xx IPU2 processor. The DSP CMA start address and size are
unchanged. This reduced size causes the CMA pools to end at
0x99000000 instead of the previous 0xA0000000.

NOTE: This requires the BIOS-side images to use the reduced carveout
entries in the resource table, so older images with different resource
tables will not boot after this change.

Signed-off-by: Suman Anna <s-anna@ti.com>